backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
nba_api==1.5.2
pbpstats==1.0.6
pandas==2.2.2
sqlalchemy==2.0.35
asyncpg==0.29.0
alembic==1.13.2
python-dotenv==1.0.1
httpx==0.27.2
pydantic==2.9.2
pydantic-settings==2.5.2
scikit-learn==1.5.2
numpy==1.26.4
anthropic>=0.40.0
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
